Solution to Find the Majority Element in an Array

+++ Find the Majority Element in an Array@@676c02002dbc8c3e7a280b76/code.c
@@ -3,7 +3,7 @@
 int main(){
     int n;
     scanf("%d",&n);
-    float arr[n];
+    int arr[n];
     for (int i=0;i<n;i++){
         scanf("%f",&arr[i]);
     }
